Move tests around
